refactor(ui/styles): dynamic 제거 enum을 utils로 변경
