+++ packages/ui/styles/src/index.ts
@@ -1,5 +1,4 @@
 /* eslint-disable @stylistic/padding-line-between-statements, react-refresh/only-export-components */
 export * from './utils';
-export * from './extract-dynamic-props';
 export * from './make-data';
 export * from './map-to-prop';
