--- packages/ui/styles/src/utils/dynamic/flex-basis.css.ts
@@ -1,7 +1,0 @@
-import { createVar, style } from '@vanilla-extract/css';
-
-export const flexBasisVar = createVar();
-
-export const flexBasis = style({
-  flexBasis: flexBasisVar,
-});
