--- packages/ui/styles/src/utils/dynamic/height.css.ts
@@ -1,7 +1,0 @@
-import { createVar, style } from '@vanilla-extract/css';
-
-export const heightVar = createVar();
-
-export const height = style({
-  height: heightVar,
-});
