--- packages/ui/styles/src/utils/dynamic/index.ts
@@ -1,11 +1,0 @@
-/* eslint-disable @stylistic/padding-line-between-statements, react-refresh/only-export-components */
-export * from './flex-basis.css';
-export * from './gap.css';
-export * from './height.css';
-export * from './inset.css';
-export * from './max-height.css';
-export * from './max-width.css';
-export * from './min-height.css';
-export * from './min-width.css';
-export * from './padding.css';
-export * from './width.css';
