--- packages/ui/styles/src/utils/dynamic/max-height.css.ts
@@ -1,7 +1,0 @@
-import { createVar, style } from '@vanilla-extract/css';
-
-export const maxHeightVar = createVar();
-
-export const maxHeight = style({
-  maxHeight: maxHeightVar,
-});
