--- packages/ui/styles/src/utils/dynamic/max-width.css.ts
@@ -1,7 +1,0 @@
-import { createVar, style } from '@vanilla-extract/css';
-
-export const maxWidthVar = createVar();
-
-export const maxWidth = style({
-  maxWidth: maxWidthVar,
-});
