--- packages/ui/styles/src/utils/dynamic/min-height.css.ts
@@ -1,7 +1,0 @@
-import { createVar, style } from '@vanilla-extract/css';
-
-export const minHeightVar = createVar();
-
-export const minHeight = style({
-  minHeight: minHeightVar,
-});
