--- packages/ui/styles/src/utils/dynamic/min-width.css.ts
@@ -1,7 +1,0 @@
-import { createVar, style } from '@vanilla-extract/css';
-
-export const minWidthVar = createVar();
-
-export const minWidth = style({
-  minWidth: minWidthVar,
-});
