--- packages/ui/styles/src/utils/dynamic/width.css.ts
@@ -1,7 +1,0 @@
-import { createVar, style } from '@vanilla-extract/css';
-
-export const widthVar = createVar();
-
-export const width = style({
-  width: widthVar,
-});
