--- packages/ui/styles/src/utils/enum/index.ts
@@ -1,15 +1,0 @@
-/* eslint-disable @stylistic/padding-line-between-statements, react-refresh/only-export-components */
-export * from './align-items.css';
-export * from './border-radius.css';
-export * from './display.css';
-export * from './flex-direction.css';
-export * from './flex-grow.css';
-export * from './flex-shrink.css';
-export * from './flex-wrap.css';
-export * from './justify-content.css';
-export * from './letter-spacing.css';
-export * from './overflow.css';
-export * from './position.css';
-export * from './text-align.css';
-export * from './text-wrap.css';
-export * from './truncate.css';
