+++ packages/ui/styles/src/utils/flex-grow.css.ts
@@ -1,6 +1,6 @@
 import { styleVariants } from '@vanilla-extract/css';
-import { makeCustomData } from '../../make-data';
-import { mapToProp } from '../../map-to-prop';
+import { makeCustomData } from '../make-data';
+import { mapToProp } from '../map-to-prop';
 
 const flexGrowKeys = ['none', 'grow'] as const;
