+++ packages/ui/styles/src/utils/justify-content.css.ts
@@ -1,6 +1,6 @@
 import { type CSSProperties, styleVariants } from '@vanilla-extract/css';
-import { makeCustomData } from '../../make-data';
-import { mapToProp } from '../../map-to-prop';
+import { makeCustomData } from '../make-data';
+import { mapToProp } from '../map-to-prop';
 
 const justifyContentKeys = ['start', 'center', 'end', 'spaceBetween'] as const;
