+++ packages/ui/styles/src/utils/text-align.css.ts
@@ -1,6 +1,6 @@
 import { type CSSProperties, styleVariants } from '@vanilla-extract/css';
-import { makeData } from '../../make-data';
-import { mapToProp } from '../../map-to-prop';
+import { makeData } from '../make-data';
+import { mapToProp } from '../map-to-prop';
 
 const textAlignValues = [
   'left',
