+++ packages/ui/styles/src/utils/truncate.css.ts
